pyproject.toml
[project]
name = "bolthackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.3",
    "django-cors-headers>=4.7.0",
    "django-drf-otp>=0.0.11",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "gunicorn>=23.0.0",
    "markdown>=3.8.2",
    "openai>=1.90.0",
    "pdf2image>=1.17.0",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
]
